backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.4
celery==5.3.6
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
boto3==1.34.101
pillow==10.3.0
openpyxl==3.1.2
python-dotenv==1.0.1
